fastapi/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
redis
mysql-connector-python
faiss-cpu
numpy
langchain-openai
tqdm
